+++ stl/lib/stl.dart
@@ -19,4 +19,4 @@ export 'src/ranges/repeat_range.dart';
 export 'src/ranges/cartesian_range.dart';
 export 'src/utility/pair.dart';
 export 'src/utility/optional.dart';
-
+export 'src/utility/variant.dart';